PAYEE-594: Route relocated to testOnlyDoNotUseInAppConf and test opeartion moved to its own test controller not to be released to PROD
